Significant update to readme and server, moderate change to client.
The client will now also exit on some server messages, namely the server
shutting down and the server not being up.

The readme has been updated with the emails (in markdown, just click on
them to view the emails) and with 3 bonus mark functions, namely the
deployment, the admin, and the login management.

The server now has functionality to output the list of courses the user
is connected to.